Split normalizers again
